Bug fixes for aggregation and object WriteTo json
