+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Query/ContinuationTokens/OrderByContinuationTokenTests.cs
@@ -57,32 +57,6 @@ namespace Microsoft.Azure.Cosmos.Query
         }
 
         [TestMethod]
-        public void TestOrderByUndefined()
-        {
-            string testResponse = @"{""_rid"":""CuECAN5Z6bM="",""Documents"":[{""_rid"":""CuECAN5Z6bMOAAAAAAAAAA=="",""orderByItems"":[{}]},{""_rid"":""CuECAN5Z6bMNAAAAAAAAAA=="",""orderByItems"":[{}]},{""_rid"":""CuECAN5Z6bMMAAAAAAAAAA=="",""orderByItems"":[{}]},{""_rid"":""CuECAN5Z6bMLAAAAAAAAAA=="",""orderByItems"":[{}]},{""_rid"":""CuECAN5Z6bMKAAAAAAAAAA=="",""orderByItems"":[{}]},{""_rid"":""CuECAN5Z6bMJAAAAAAAAAA=="",""orderByItems"":[{}]},{""_rid"":""CuECAN5Z6bMIAAAAAAAAAA=="",""orderByItems"":[{}]},{""_rid"":""CuECAN5Z6bMHAAAAAAAAAA=="",""orderByItems"":[{}]},{""_rid"":""CuECAN5Z6bMGAAAAAAAAAA=="",""orderByItems"":[{}]},{""_rid"":""CuECAN5Z6bMFAAAAAAAAAA=="",""orderByItems"":[{}]},{""_rid"":""CuECAN5Z6bMEAAAAAAAAAA=="",""orderByItems"":[{}]},{""_rid"":""CuECAN5Z6bMDAAAAAAAAAA=="",""orderByItems"":[{}]},{""_rid"":""CuECAN5Z6bMCAAAAAAAAAA=="",""orderByItems"":[{}]},{""_rid"":""CuECAN5Z6bMBAAAAAAAAAA=="",""orderByItems"":[{}]}],""_count"":14}";
-
-            MemoryStream memoryStream = new MemoryStream(Encoding.UTF8.GetBytes(testResponse));
-
-            CosmosArray documents = CosmosQueryClientCore.ParseElementsFromRestStream(
-                memoryStream,
-                Documents.ResourceType.Document,
-                new CosmosSerializationFormatOptions(
-                    "JsonText",
-                    (content) => JsonNavigator.Create(content),
-                    () => JsonWriter.Create(JsonSerializationFormat.Text)));
-
-            List<OrderByQueryResult> orderByQueryResults = documents.Select(x => new OrderByQueryResult(x)).ToList();
-            Assert.AreEqual(14, orderByQueryResults.Count);
-
-            foreach (OrderByQueryResult orderByQueryResult in orderByQueryResults)
-            {
-                Assert.IsTrue(orderByQueryResult.Payload.Equals(CosmosUndefined.Instance));
-                Assert.AreEqual(1, orderByQueryResult.OrderByItems.Count());
-                Assert.IsTrue(orderByQueryResult.OrderByItems.First().Item.Equals(CosmosUndefined.Instance));
-            }
-        }
-
-        [TestMethod]
         public void TestOrderByQueryLiterals()
         {
             StringBuilder sb = new StringBuilder();
